+++ backend-api/CallCenterAPI_FastAPI/schemas.py
@@ -259,7 +259,7 @@ class UserOut(BaseModel):
     managerId: Optional[int] = None
     phone: Optional[str] = None
     fatherName: Optional[str] = None
-    monthlySalary: Optional[float] = 0.0
+    monthlySalary: Optional[int] = 0
     cnic: Optional[str] = None
     department: Optional[str] = None
     designation: Optional[str] = None
@@ -279,7 +279,7 @@ class AgentOut(BaseModel):
     managerId: Optional[int] = None
     phone: Optional[str] = None
     fatherName: Optional[str] = None
-    monthlySalary: Optional[float] = 0.0
+    monthlySalary: Optional[int] = 0
     cnic: Optional[str] = None
     department: Optional[str] = None
     designation: Optional[str] = None
@@ -328,7 +328,7 @@ class CreateAgentRequest(BaseModel):
     password: str
     managerId: int
     fatherName: Optional[str] = None
-    monthlySalary: Optional[float] = 0.0
+    monthlySalary: Optional[int] = 0
     cnic: Optional[str] = None
     phone: Optional[str] = None
     department: Optional[str] = None
@@ -356,7 +356,7 @@ class UpdateAgentStaffRequest(BaseModel):
     email: Optional[str] = None
     phone: Optional[str] = None
     fatherName: Optional[str] = None
-    monthlySalary: Optional[float] = 0.0
+    monthlySalary: Optional[int] = 0
     cnic: Optional[str] = None
     department: Optional[str] = None
     designation: Optional[str] = None
